+++ sdk/src/main/java/com/hedera/hashgraph/sdk/ContractUpdateTransaction.java
@@ -19,6 +19,7 @@
  */
 package com.hedera.hashgraph.sdk;
 
+import com.google.protobuf.Int32Value;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.StringValue;
 import com.hedera.hashgraph.sdk.proto.ContractUpdateTransactionBody;
@@ -27,6 +28,7 @@ import com.hedera.hashgraph.sdk.proto.SmartContractServiceGrpc;
 import com.hedera.hashgraph.sdk.proto.TransactionBody;
 import com.hedera.hashgraph.sdk.proto.TransactionResponse;
 import io.grpc.MethodDescriptor;
+import org.checkerframework.checker.units.qual.A;
 import org.threeten.bp.Duration;
 import org.threeten.bp.Instant;
 
@@ -68,6 +70,10 @@ public final class ContractUpdateTransaction extends Transaction<ContractUpdateT
     private Instant expirationTime = null;
     @Nullable
     private Key adminKey = null;
+    @Nullable
+    private Integer maxAutomaticTokenAssociations = null;
+    @Nullable
+    private AccountId autoRenewAccountId = null;
     @Nullable
     private Duration autoRenewPeriod = null;
     @Nullable
@@ -167,6 +173,52 @@ public final class ContractUpdateTransaction extends Transaction<ContractUpdateT
     }
 
     @Nullable
+    public Integer getMaxAutomaticTokenAssociations() {
+        return maxAutomaticTokenAssociations;
+    }
+
+    /**
+     * Sets the new maximum number of tokens that this contract can be
+     * automatically associated with (i.e., receive air-drops from).
+     *
+     * @param maxAutomaticTokenAssociations The maximum automatic token associations
+     * @return  {@code this}
+     */
+
+    public ContractUpdateTransaction setMaxAutomaticTokenAssociations(int maxAutomaticTokenAssociations) {
+        requireNotFrozen();
+        this.maxAutomaticTokenAssociations = maxAutomaticTokenAssociations;
+        return this;
+    }
+
+    @Nullable
+    public AccountId getAutoRenewAccountId() {
+        return autoRenewAccountId;
+    }
+
+    /**
+     * Sets account to charge for auto-renewal of this contract. If not set, or set to an
+     * account with zero hbar balance, the contract's own hbar balance will be used to
+     * cover auto-renewal fees.
+     *
+     * @param accountId ID of account to charge fees to
+     * @return {@code this}
+     */
+
+    public ContractUpdateTransaction setAutoRenewAccountId(AccountId accountId) {
+        Objects.requireNonNull(accountId);
+        requireNotFrozen();
+        autoRenewAccountId = accountId;
+        return this;
+    }
+
+    public ContractUpdateTransaction clearAutoRenewAccountId() {
+        requireNotFrozen();
+        autoRenewAccountId = new AccountId(0, 0, 0);
+        return this;
+    }
+
+    @Nullable
     public Duration getAutoRenewPeriod() {
         return autoRenewPeriod;
     }
@@ -250,6 +302,12 @@ public final class ContractUpdateTransaction extends Transaction<ContractUpdateT
         if (body.hasAdminKey()) {
             adminKey = Key.fromProtobufKey(body.getAdminKey());
         }
+        if (body.hasMaxAutomaticTokenAssociations()) {
+            maxAutomaticTokenAssociations = body.getMaxAutomaticTokenAssociations().getValue();
+        }
+        if (body.hasAutoRenewAccountId()) {
+            autoRenewAccountId = AccountId.fromProtobuf(body.getAutoRenewAccountId());
+        }
         if (body.hasAutoRenewPeriod()) {
             autoRenewPeriod = DurationConverter.fromProtobuf(body.getAutoRenewPeriod());
         }
@@ -272,6 +330,12 @@ public final class ContractUpdateTransaction extends Transaction<ContractUpdateT
         if (adminKey != null) {
             builder.setAdminKey(adminKey.toProtobufKey());
         }
+        if (maxAutomaticTokenAssociations != null) {
+            builder.setMaxAutomaticTokenAssociations(Int32Value.of(maxAutomaticTokenAssociations));
+        }
+        if (autoRenewAccountId != null) {
+            builder.setAutoRenewAccountId(autoRenewAccountId.toProtobuf());
+        }
         if (autoRenewPeriod != null) {
             builder.setAutoRenewPeriod(DurationConverter.toProtobuf(autoRenewPeriod));
         }
@@ -289,6 +353,10 @@ public final class ContractUpdateTransaction extends Transaction<ContractUpdateT
 
         if (proxyAccountId != null) {
             proxyAccountId.validateChecksum(client);
+        }
+
+        if (autoRenewAccountId != null) {
+            autoRenewAccountId.validateChecksum(client);
         }
     }