+++ sdk/src/test/java/com/hedera/hashgraph/sdk/ContractCreateTransactionTest.java
@@ -20,6 +20,7 @@
 package com.hedera.hashgraph.sdk;
 
 import io.github.jsonSnapshot.SnapshotMatcher;
+import org.bouncycastle.util.encoders.Hex;
 import org.junit.AfterClass;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -53,6 +54,13 @@ public class ContractCreateTransactionTest {
         ).toMatchSnapshot();
     }
 
+    @Test
+    void shouldSerialize2() {
+        SnapshotMatcher.expect(spawnTestTransaction2()
+            .toString()
+        ).toMatchSnapshot();
+    }
+
     private ContractCreateTransaction spawnTestTransaction() {
         return new ContractCreateTransaction()
             .setNodeAccountIds(Arrays.asList(AccountId.fromString("0.0.5005"), AccountId.fromString("0.0.5006")))
@@ -62,6 +70,26 @@ public class ContractCreateTransactionTest {
             .setGas(0)
             .setInitialBalance(Hbar.fromTinybars(1000))
             .setProxyAccountId(AccountId.fromString("0.0.1001"))
+            .setMaxAutomaticTokenAssociations(101)
+            .setAutoRenewAccountId(AccountId.fromString("0.0.456"))
+            .setAutoRenewPeriod(Duration.ofHours(10))
+            .setConstructorParameters(new byte[]{10, 11, 12, 13, 25})
+            .setMaxTransactionFee(Hbar.fromTinybars(100_000))
+            .freeze()
+            .sign(unusedPrivateKey);
+    }
+
+    private ContractCreateTransaction spawnTestTransaction2() {
+        return new ContractCreateTransaction()
+            .setNodeAccountIds(Arrays.asList(AccountId.fromString("0.0.5005"), AccountId.fromString("0.0.5006")))
+            .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
+            .setBytecode(Hex.decode("deadbeef"))
+            .setAdminKey(unusedPrivateKey)
+            .setGas(0)
+            .setInitialBalance(Hbar.fromTinybars(1000))
+            .setProxyAccountId(AccountId.fromString("0.0.1001"))
+            .setMaxAutomaticTokenAssociations(101)
+            .setAutoRenewAccountId(AccountId.fromString("0.0.456"))
             .setAutoRenewPeriod(Duration.ofHours(10))
             .setConstructorParameters(new byte[]{10, 11, 12, 13, 25})
             .setMaxTransactionFee(Hbar.fromTinybars(100_000))
@@ -74,5 +102,12 @@ public class ContractCreateTransactionTest {
         var tx = spawnTestTransaction();
         var tx2 = ContractCreateTransaction.fromBytes(tx.toBytes());
         assertThat(tx2.toString()).isEqualTo(tx.toString());
+    }
+
+    @Test
+    void shouldBytes2() throws Exception {
+        var tx = spawnTestTransaction2();
+        var tx2 = ContractCreateTransaction.fromBytes(tx.toBytes());
+        assertEquals(tx.toString(), tx2.toString());
     }
 }