+++ sdk/src/test/java/com/hedera/hashgraph/sdk/ContractFunctionResultTest.java
@@ -74,7 +74,8 @@ public class ContractFunctionResultTest {
                                 BigInteger.valueOf(777)
                             )
                         )
-                    ).toProtobuf())
+                    ).toProtobuf()
+                ).setSenderId(AccountId.fromString("1.2.3").toProtobuf())
         );
 
         // interpretation varies based on width
@@ -95,6 +96,8 @@ public class ContractFunctionResultTest {
 
         assertThat(result.getString(4)).isEqualTo("Hello, world!");
         assertThat(result.getString(5)).isEqualTo("Hello, world, again!");
+
+        assertThat(result.senderAccountId).isEqualTo(AccountId.fromString("1.2.3"));
 
         assertThat(result.contractId).isEqualTo(ContractId.fromString("1.2.3"));
         assertThat(result.evmAddress).isEqualTo(ContractId.fromEvmAddress(1, 2, "98329e006610472e6B372C080833f6D79ED833cf"));