+++ sdk/src/test/java/com/hedera/hashgraph/sdk/ContractUpdateTransactionTest.java
@@ -59,6 +59,8 @@ public class ContractUpdateTransactionTest {
             .setTransactionId(TransactionId.withValidStart(AccountId.fromString("0.0.5006"), validStart))
             .setContractId(ContractId.fromString("0.0.5007"))
             .setAdminKey(privateKey)
+            .setMaxAutomaticTokenAssociations(101)
+            .setAutoRenewAccountId(AccountId.fromString("0.0.456"))
             .setAutoRenewPeriod(Duration.ofDays(1))
             .setContractMemo("3")
             .setExpirationTime(Instant.ofEpochMilli(4))